refactor: extract TaskScheduler into own library
