+++ packages/zwave-js/package.json
@@ -139,6 +139,7 @@
     "@zwave-js/serial": "workspace:*",
     "@zwave-js/shared": "workspace:*",
     "@zwave-js/testing": "workspace:*",
+    "@zwave-js/waddle": "^1.0.0",
     "alcalzone-shared": "^5.0.0",
     "ansi-colors": "^4.1.3",
     "ky": "^1.7.2",
